+++ src/game/tests.rs
@@ -432,6 +432,7 @@ mod vintage_cube_grave_titan;
 mod vintage_cube_graveyard;
 mod vintage_cube_green_suns_zenith;
 mod vintage_cube_grief;
+mod vintage_cube_grist_the_hunger_tide;
 mod vintage_cube_guide_of_souls;
 mod vintage_cube_gut;
 mod vintage_cube_hand_attack;